requirements.txt
Flask==1.1.2
PyYAML==5.3.1
requests==2.25.1
waitress==1.4.4
